adding page for upload dar,spl and refactoring code(remove unneccassry code
